+++ src/app/layout.tsx
@@ -32,14 +32,15 @@ export default function RootLayout({
     <html lang="ko" className={`${pretendard.variable}`}>
       <body>
         <QueryProvider>
-          <AuthTokenRefreshProvider />
-          <AuthSessionSync />
-          {children}
-          <Script
-            src="//t1.kakaocdn.net/mapjsapi/bundle/postcode/prod/postcode.v2.js"
-            strategy="lazyOnload"
-          />
-          <ToastContainer />
+          <AuthTokenRefreshProvider>
+            <AuthSessionSync />
+            {children}
+            <Script
+              src="//t1.kakaocdn.net/mapjsapi/bundle/postcode/prod/postcode.v2.js"
+              strategy="lazyOnload"
+            />
+            <ToastContainer />
+          </AuthTokenRefreshProvider>
         </QueryProvider>
         <OverlayRoot />
         <div id={OVERLAY_ROOT_ID} />
